+++ src/deletion/basecache.ts
